+++ hector_gamepad_manager_interface/README.md
@@ -1,0 +1,38 @@
+# hector_gamepad_manager_interface
+
+Provides plugin base class for gamepad manager plugins
+
+- [hector_gamepad_manager_interface](#hector_gamepad_manager_interface)
+
+
+## `hector_gamepad_manager_interface`
+
+### Subscribed Topics
+
+| Topic | Type | Description |
+| --- | --- | --- |
+|  |  |  |
+
+### Published Topics
+
+| Topic | Type | Description |
+| --- | --- | --- |
+|  |  |  |
+
+### Services
+
+| Service | Type | Description |
+| --- | --- | --- |
+|  |  |  |
+
+### Actions
+
+| Action | Type | Description |
+| --- | --- | --- |
+|  |  |  |
+
+### Parameters
+
+| Parameter | Type | Description |
+| --- | --- | --- |
+|  |  |  |
