+++ hector_gamepad_manager_interface/package.xml
@@ -1,0 +1,19 @@
+<?xml version="1.0"?>
+<?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
+<package format="3">
+  <name>hector_gamepad_manager_interface</name>
+  <version>0.0.0</version>
+  <description>Provides plugin base class for gamepad manager plugins</description>
+
+  <maintainer email="aljoscha.schmidt@tu-darmstadt.de">Aljoscha Schmidt</maintainer>
+  <license>TODO</license>
+  <author email="aljoscha.schmidt@tu-darmstadt.de">Aljoscha Schmidt</author>
+
+  <buildtool_depend>ament_cmake</buildtool_depend>
+  <depend>rclcpp</depend>
+  <depend>std_msgs</depend>
+
+  <export>
+    <build_type>ament_cmake</build_type>
+  </export>
+</package>
