+++ hector_gamepad_manager_plugins/include/hector_gamepad_manager_plugins/flipper_plugin.hpp
@@ -1,7 +1,7 @@
 #ifndef HECTOR_GAMEPAD_MANAGER_PLUGINS_FLIPPER_PLUGIN_HPP
 #define HECTOR_GAMEPAD_MANAGER_PLUGINS_FLIPPER_PLUGIN_HPP
 
-#include <hector_gamepad_manager/gamepad_function_plugin.hpp>
+#include <hector_gamepad_manager_interface/gamepad_plugin_interface.hpp>
 
 #include "hector_gamepad_manager_plugins/controller_helper.hpp"
 #include <hector_ros2_utils/parameters/reconfigurable_parameter.hpp>
@@ -9,7 +9,7 @@
 
 namespace hector_gamepad_manager_plugins
 {
-class FlipperPlugin : public hector_gamepad_manager::GamepadFunctionPlugin
+class FlipperPlugin : public hector_gamepad_manager_interface::GamepadFunctionPlugin
 {
 public:
   void initialize( const rclcpp::Node::SharedPtr &node ) override;
