+++ hector_gamepad_manager_plugins/include/hector_gamepad_manager_plugins/manipulation_plugin.hpp
@@ -4,7 +4,7 @@
 #ifndef HECTOR_GAMEPAD_MANAGER_PLUGINS_MANIPULATION_PLUGIN_HPP
 #define HECTOR_GAMEPAD_MANAGER_PLUGINS_MANIPULATION_PLUGIN_HPP
 
-#include <hector_gamepad_manager/gamepad_function_plugin.hpp>
+#include <hector_gamepad_manager_interface/gamepad_plugin_interface.hpp>
 #include <hector_gamepad_manager_plugins/controller_helper.hpp>
 #include <hector_ros2_utils/parameters/reconfigurable_parameter.hpp>
 #include <geometry_msgs/msg/twist_stamped.hpp>
@@ -14,7 +14,7 @@
 
 namespace hector_gamepad_manager_plugins
 {
-class ManipulationPlugin final : public hector_gamepad_manager::GamepadFunctionPlugin
+class ManipulationPlugin final : public hector_gamepad_manager_interface::GamepadFunctionPlugin
 {
 public:
   void initialize( const rclcpp::Node::SharedPtr &node ) override;
