+++ hector_gamepad_manager_plugins/src/flipper_plugin.cpp
@@ -249,4 +249,4 @@ bool FlipperPlugin::checkCurrentCmdIsZero() const
 #include <pluginlib/class_list_macros.hpp>
 
 PLUGINLIB_EXPORT_CLASS( hector_gamepad_manager_plugins::FlipperPlugin,
-                        hector_gamepad_manager::GamepadFunctionPlugin )
+                        hector_gamepad_manager_interface::GamepadFunctionPlugin )
