+++ hector_gamepad_manager_plugins/src/manipulation_plugin.cpp
@@ -215,4 +215,4 @@ bool ManipulationPlugin::isZeroCmd() const
 #include <pluginlib/class_list_macros.hpp>
 
 PLUGINLIB_EXPORT_CLASS( hector_gamepad_manager_plugins::ManipulationPlugin,
-                        hector_gamepad_manager::GamepadFunctionPlugin )
+                        hector_gamepad_manager_interface::GamepadFunctionPlugin )
